requirements.txt
pydantic==2.12.5
openai==2.30.0
fastapi==0.135.2
uvicorn==0.42.0
pyyaml==6.0.1
